Add lzss.s and hasm support
