+++ config/symbols.main.us.txt
@@ -279,7 +279,8 @@ SysCdromLoadLzs = 0x80033FC4;
 SystemCdromAbortLoading = 0x8003408C;
 SystemCdromReadChain = 0x80034B44;
 SystemLzsDecompress = 0x80034BB0;
-ChangeClearSIO = 0x80034CE8;
+D_80034CF0 = 0x80034CF0; // size:0x24
+D_80034D14 = 0x80034D14; // size:0x4
 SysCdromGetPackPointer = 0x80034D18;
 SysCdromSetLzsExtract = 0x80034D2C;
 SysCdromLzsExtract = 0x80034E00;
