+++ config/us.yaml
@@ -45,7 +45,9 @@ overlays:
       - [0x171C0, c, 269C0]
       - [0x1735C, c, 26B70]
       - [0x19BD0, c, akao]
-      - [0x24370, c, psxsdk]
+      - [0x24370, c, 33B70]
+      - [0x254F0, hasm, lzss]
+      - [0x2573C, c, psxsdk]
       - [0x394E0, .data, psxsdk/card_data]
       - [0x394FC, .data, 110B8]
       - [0x39584, .data, 144D8]
